+++ pkg/providers/property/testdata/TestDataReportingGroupsGroups/reporting_groups_with_contract_id.tf
@@ -1,0 +1,8 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+data "akamai_property_reportinggroups_groups" "test" {
+  contract_id = "ctr_456"
+}
+