+++ pkg/providers/property/testdata/TestDataReportingGroupsGroups/reporting_groups_with_cp_code_id.tf
@@ -1,0 +1,8 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+data "akamai_property_reportinggroups_groups" "test" {
+  cp_code_id = "111"
+}
+